requirements-dev.txt
-r requirements.txt
django-browser-reload==1.21.0
django-watchfiles==1.4.0
